chore(rust-api): extract server.rs (Settings/AppState/run/run_worker) + explicit imports (sc-9736)
Deferred remainder of sc-8890 (F-088, the rust-api lib.rs grab-bag). Pure
behavior-preserving code-move + import tidy; no logic, API, or contract change.

1. Move `Settings` (struct + impl), `AppState`, `run`, and `run_worker` out of
   the crate root into a dedicated `server.rs` module and re-export them from
   lib.rs (`pub use server::{run, run_worker, AppState, Settings}`) so every
   external reference — `main.rs` (`run`/`run_worker`), the handler modules'
   `use super::*` (`AppState`), and `tests.rs` (`Settings`) — keeps resolving
   unchanged. `AppState` fields become `pub(crate)` so the crate-root
   `create_app_with_state` literal and the sibling handlers still access them.
   The startup helpers (`create_app`, `spawn_inprocess_utility_worker`,
   `shutdown_signal`, `should_warn_open_bind`, `open_bind_override_enabled`,
   `parent_death`, `parent_pid_to_watch`, `env_string`, `env_path_or`) and the
   config constants stay in lib.rs; server.rs reaches them via `crate::`.

2. Replace the 23 `use module::*` glob imports in lib.rs with explicit,
   named imports (F-088's readability goal). No wildcard `use x::*` remains
   except the deliberate `pub(crate) use defaults::*` serde-provider prelude
   re-export from sc-8890 (its ~40 `#[serde(default = "default_x")]` string
   paths resolve through the DTO modules' `use super::*`). A handful of helpers
   consumed only by the `tests` module via `super::` are brought in through
   `#[cfg(test)]`-gated crate-root imports so the non-test build stays
   warning-free.

Verified: `cargo build`/`test` (185 pass) / `clippy --all-targets -D warnings`
green; `cargo fmt --all -- --check` clean; the rust-api contract snapshots
(`tests/fixtures/rust_api_contract_snapshots/snapshots.json`) regenerate with
no diff, confirming the contract is unchanged.
